requirements.txt
attrs==17.2.0
certifi==2017.4.17
chardet==3.0.3
click==6.7
Flask==0.12.2
Flask-OAuthlib==0.9.3
future==0.16.0
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
oauthlib==2.0.2
python-twitter==3.3
requests==2.17.3
requests-oauthlib==0.8.0
urllib3==1.21.1
Werkzeug==0.12.2
